Reorganizes manifest files
Manifests will now be always be placed in the "manifests" directory,
and any given build will always have the same filename. Each release
will have a tag, allowing access to the manifests for a specific
release. This allows the diff for each release to display the changes
in the manifests between releases.

Manifest files in this commit are from the 2017.06.1 release.