Pauses briefly in attempt to ensure inline script exists
Periodically, packer would encounter the error, "No such file or
directory," when attempting to execute an inline script that packer
uploads to the system. Assumption here is that there is a race
condition where packer attempts to execute the script before the
script has been created, and that `pause_before` will cause packer
to upload the script but wait before executing it.