Uses chroot to execute the rpm manifest query
Beginning in the Feb builds, the rpm manifest was empty for the el7
builds. The exact cause is currently unknown. This patch is an
alternative approach to retrieving the manifest. Expectation is for
this to either work, or error on failure (where the prior method
appeared to be silently failing).