Moves error handling into build.sh script
Because packer can run multiple builders at a time, we do not want
to exit immediately if only one of the bulders fails. We want the
tests to run on the successful builders. However, we do still want
to exit the job with an error at the end.

This patch tracks the build exit code and the test exit code, and
handles failures in either accordingly.